requirements.txt
configobj
influxdb
python-nest
